tests: cast division result to int
Otherwise it is a float on Python 3 and code later compares about
casting a float to an int.

Differential Revision: https://phab.mercurial-scm.org/D5759

+++ tests/test-remotefilelog-datapack.py
@@ -292,7 +292,7 @@ class datapacktestsbase(object):
 
         class testdatapackstore(datapack.datapackstore):
             # Ensures that we are not keeping everything in the cache.
-            DEFAULTCACHESIZE = numpacks / 2
+            DEFAULTCACHESIZE = int(numpacks / 2)
 
         store = testdatapackstore(uimod.ui(), packdir)
